change all to real*8

+++ Components/GEOSldas_GridComp/@GEOSgcm_GridComp/GEOSagcm_GridComp/GEOSphysics_GridComp/GEOSsurface_GridComp/GEOSland_GridComp/GEOSroute_GridComp/GEOS_RouteGridComp.F90
@@ -682,7 +682,7 @@ contains
 ! ----------------------------------------------------- 
 
     real*8, dimension(:), pointer :: RUNOFF 
-    real*8, dimension(:), pointer :: RUNOFF_SRC0   
+    real, dimension(:), pointer :: RUNOFF_SRC0   
 
 ! -----------------------------------------------------
 ! INTERNAL pointers
